+++ src/Level04/Main.hs
@@ -54,7 +54,6 @@ runApp = error "runApp needs re-implementing"
 --
 -- 1) Load the configuration.
 -- 2) Attempt to initialise the database.
--- 3) Combine the results into a tuple
 --
 -- Our application configuration is defined in Conf.hs
 --
