+++ src/Level05/DB.hs
@@ -9,6 +9,8 @@ module Level05.DB
   , deleteTopic
   ) where
 
+import           Control.Monad.IO.Class             (liftIO)
+
 import           Data.Text                          (Text)
 import qualified Data.Text                          as Text
 
@@ -22,11 +24,13 @@ import qualified Database.SQLite.Simple             as Sql
 import qualified Database.SQLite.SimpleErrors       as Sql
 import           Database.SQLite.SimpleErrors.Types (SQLiteResponse)
 
-import           Level05.Types                     (Comment, CommentText,
+import           Level05.Types                      (Comment, CommentText,
                                                      Error (DBError), Topic,
                                                      fromDbComment,
                                                      getCommentText, getTopic,
                                                      mkTopic)
+
+import           Level05.AppM                       (AppM)
 
 -- We have a data type to simplify passing around the information we need to run
 -- our database queries. This also allows things to change over time without
@@ -64,64 +68,34 @@ initDB fp = Sql.runDBAction $ do
 runDB
   :: (a -> Either Error b)
   -> IO a
-  -> IO (Either Error b)
-runDB f a =
-  (>>= f) . first DBError <$> Sql.runDBAction a
-  -- Choices, choices...
-  -- either (Left . DBError) f <$> Sql.runDBAction a
-  -- these two are pretty much the same.
-  -- Sql.runDBAction >=> pure . either (Left . DBError) f
-  -- this is because we noticed that our call to pure, which means we should
-  -- be able to fmap to victory.
-  -- fmap ( either (Left . DBError) f ) . Sql.runDBAction
+  -> AppM b
+runDB =
+  error "Copy your completed 'runDB' and refactor to match the new type signature"
 
 getComments
   :: FirstAppDB
   -> Topic
-  -> IO (Either Error [Comment])
-getComments db t = do
-  -- Write the query with an icky string and remember your placeholders!
-  let q = "SELECT id,topic,comment,time FROM comments WHERE topic = ?"
-  -- To be doubly and triply sure we've no garbage in our response, we take care
-  -- to convert our DB storage type into something we're going to share with the
-  -- outside world. Checking again for things like empty Topic or CommentText values.
-  runDB (traverse fromDbComment) $ Sql.query (dbConn db) q (Sql.Only . getTopic $ t)
+  -> AppM [Comment]
+getComments =
+  error "Copy your completed 'getComments' and refactor to match the new type signature"
 
 addCommentToTopic
   :: FirstAppDB
   -> Topic
   -> CommentText
-  -> IO (Either Error ())
-addCommentToTopic db t c = do
-  -- Record the time this comment was created.
-  nowish <- getCurrentTime
-  -- Note the triple, matching the number of values we're trying to insert, plus
-  -- one for the table name.
-  let q =
-        -- Remember that the '?' are order dependent so if you get your input
-        -- parameters in the wrong order, the types won't save you here. More on that
-        -- sort of goodness later.
-        "INSERT INTO comments (topic,comment,time) VALUES (?,?,?)"
-  -- We use the execute function this time as we don't care about anything
-  -- that is returned. The execute function will still return the number of rows
-  -- affected by the query, which in our case should always be 1.
-  runDB Right $ Sql.execute (dbConn db) q (getTopic t, getCommentText c, nowish)
-  -- An alternative is to write a returning query to get the Id of the DbComment
-  -- we've created. We're being lazy (hah!) for now, so assume awesome and move on.
+  -> AppM ()
+addCommentToTopic =
+  error "Copy your completed 'appCommentToTopic' and refactor to match the new type signature"
 
 getTopics
   :: FirstAppDB
-  -> IO (Either Error [Topic])
-getTopics db =
-  let q = "SELECT DISTINCT topic FROM comments"
-  in
-    runDB (traverse ( mkTopic . Sql.fromOnly )) $ Sql.query_ (dbConn db) q
+  -> AppM [Topic]
+getTopics =
+  error "Copy your completed 'getTopics' and refactor to match the new type signature"
 
 deleteTopic
   :: FirstAppDB
   -> Topic
-  -> IO (Either Error ())
-deleteTopic db t =
-  let q = "DELETE FROM comments WHERE topic = ?"
-  in
-    runDB Right $ Sql.execute (dbConn db) q (Sql.Only . getTopic $ t)
+  -> AppM ()
+deleteTopic =
+  error "Copy your completed 'deleteTopic' and refactor to match the new type signature"
