+++ src/Level05/Types.hs
@@ -4,11 +4,6 @@
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 module Level05.Types
   ( Error (..)
-  , ConfigError (..)
-  , PartialConf (..)
-  , Port (..)
-  , DBFilePath (..)
-  , Conf (..)
   , RqType (..)
   , ContentType (..)
   , Comment (..)
@@ -19,7 +14,6 @@ module Level05.Types
   , mkCommentText
   , getCommentText
   , renderContentType
-  , confPortToWai
   , fromDbComment
   ) where
 
@@ -140,105 +134,3 @@ renderContentType
   -> ByteString
 renderContentType PlainText = "text/plain"
 renderContentType JSON      = "application/json"
-
------------------
--- Config Types
------------------
-
--- This is an alternative way of defining a `newtype`. You define it as a record
--- with a single field, this provides the unwrapping function for free. When
--- defined using the other method, you must use pattern-matching or write a dedicated
--- function in order to get the value out.
---
-newtype Port = Port
-  -- You will notice we're using ``Word16`` as our type for the ``Port`` value.
-  -- This is because a valid port number can only be a 16bit unsigned integer.
-  { getPort :: Word16 }
-  deriving (Eq, Show)
-
-newtype DBFilePath = DBFilePath
-  { getDBFilePath :: FilePath }
-  deriving (Eq, Show)
-
--- Add some fields to the ``Conf`` type:
--- - A customisable port number: ``Port``
--- - A filepath for our SQLite database: ``DBFilePath``
-data Conf = Conf
-
--- We're storing our Port as a Word16 to be more precise and prevent invalid
--- values from being used in our application. However Wai is not so stringent.
--- To accommodate this and make our lives a bit easier, we will write this
--- helper function to take ``Conf`` value and convert it to an ``Int``.
---
--- We'll need to use a function called; ``fromIntegral``, to convert our
--- ``Word16`` to an ``Int``. The type of this function is:
---
--- fromIntegral :: (Num b, Integral a) => a -> b
---
-confPortToWai
-  :: Conf
-  -> Int
-confPortToWai =
-  error "confPortToWai not implemented"
-
--- Similar to when we were considering our application types, leave this empty
--- for now and add to it as you go.
-data ConfigError = ConfigError
-  deriving Show
-
--- Our application will be able to load configuration from both a file and
--- command line input. We want to be able to use the command line to temporarily
--- override the configuration from our file. How do we combine the different
--- inputs to enable this property?
-
--- We want the command line configuration to take precedence over the File
--- configuration, so if we think about combining each of our ``Conf`` records,
--- we want to be able to write something like this:
-
--- ``defaults <> file <> commandLine``
-
--- We can use the ``Monoid`` typeclass to handle combining the ``Conf`` records
--- together, and the ``Last`` type to wrap up our values to handle the desired
--- precedence. The ``Last`` type is a wrapper for Maybe that when used with its
--- ``Monoid`` instance will always preference the last ``Just`` value that it
--- has:
-
--- Last (Just 3) <> Last (Just 1) = Last (Just 1)
--- Last Nothing  <> Last (Just 1) = Last (Just 1)
--- Last (Just 1) <> Last Nothing  = Last (Just 1)
-
--- To make this easier, we'll make a new type ``PartialConf`` that will have our
--- ``Last`` wrapped values. We can then define a ``Monoid`` instance for it and
--- have our ``Conf`` be a known good configuration.
-data PartialConf = PartialConf
-  { pcPort       :: Last Port
-  , pcDBFilePath :: Last DBFilePath
-  }
-
--- Before we can define our ``Monoid`` instance for ``PartialConf``, we'll have
--- to define a Semigroup instance. We define our ``(<>)`` function to lean
--- on the ``Semigroup`` instance for Last to always get the last value.
-instance Semigroup PartialConf where
-  _a <> _b = PartialConf
-    { pcPort       = error "pcPort (<>) not implemented"
-    , pcDBFilePath = error "pcDBFilePath (<>) not implemented"
-    }
-
--- We now define our ``Monoid`` instance for ``PartialConf``. Allowing us to
--- define our always empty configuration, which would always fail our
--- requirements. We just define `mappend` to be an alias of ``(<>)``
-instance Monoid PartialConf where
-  mempty = PartialConf mempty mempty
-  mappend = (<>)
-
--- When it comes to reading the configuration options from the command-line, we
--- use the 'optparse-applicative' package. This part of the exercise has already
--- been completed for you, feel free to have a look through the 'CommandLine'
--- module and see how it works.
---
--- For reading the configuration from the file, we're going to use the aeson
--- library to handle the parsing and decoding for us. In order to do this, we
--- have to tell aeson how to go about converting the JSON into our PartialConf
--- data structure.
-instance FromJSON PartialConf where
-  parseJSON = error "parseJSON for PartialConf not implemented yet."
