+++ src/Level05/Types/Topic.hs
@@ -1,11 +1,14 @@
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
+module Level05.Types.Topic
+  ( Topic
+  , mkTopic
+  , getTopic
+  ) where
 
-module Level05.Types.Topic (Topic, mkTopic, getTopic) where
+import           Data.Aeson          (ToJSON)
+import           Data.Text           (Text)
 
 import           Level05.Types.Error (Error (EmptyTopic), nonEmptyText)
-
-import           Data.Aeson           (ToJSON)
-import           Data.Text            (Text)
 
 newtype Topic = Topic Text
   deriving (Show, ToJSON)
