--- src/Level05/appconfig.json
@@ -1,3 +1,0 @@
-{
-  "dbFileName": "app_db.db"
-}
