--- src/Level05/test.json
@@ -1,3 +1,0 @@
-{
-  "foo": 33
-}
