+++ src/Level06/Conf.hs
@@ -7,51 +7,48 @@ module Level06.Conf
 import           GHC.Word                  (Word16)
 
 import           Data.Bifunctor            (first)
-import           Data.Monoid               (Last (..), (<>))
+import           Data.Monoid               ((<>))
 
-import           Level06.Types            (Conf (..), ConfigError (..),
+import           Level06.Types            (Conf, ConfigError,
                                             DBFilePath (DBFilePath),
-                                            PartialConf (..), Port (Port))
+                                            PartialConf, Port (Port))
 
 import           Level06.Conf.CommandLine (commandLineParser)
 import           Level06.Conf.File        (parseJSONConfigFile)
 
--- We have some sane defaults that we can always rely on, so define them using
--- our PartialConf.
+-- For the purposes of this application we will encode some default values to
+-- ensure that our application continues to function in the event of missing
+-- configuration values from either the file or command line inputs.
 defaultConf
   :: PartialConf
-defaultConf = PartialConf
-  (pure (Port 3000))
-  (pure (DBFilePath "app_db.db"))
+defaultConf =
+  error "defaultConf not implemented"
 
 -- We need something that will take our PartialConf and see if can finally build
--- a complete Conf record. Also we need to highlight any missing config values
--- by providing the relevant error.
+-- a complete ``Conf`` record. Also we need to highlight any missing values by
+-- providing the relevant error.
 makeConfig
   :: PartialConf
   -> Either ConfigError Conf
-makeConfig pc = Conf
-  <$> lastToEither MissingPort pcPort
-  <*> lastToEither MissingDBFilePath pcDBFilePath
-  where
-    -- You don't need to provide type signatures for most functions in where/let
-    -- sections. Sometimes the compiler might need a bit of help, or you would
-    -- like to be explicit in your intentions.
-    lastToEither
-      :: ConfigError
-      -> (PartialConf -> Last b)
-      -> Either ConfigError b
-    lastToEither e g =
-      (maybe (Left e) Right . getLast . g) pc
+makeConfig =
+  error "makeConfig not implemented"
 
 -- This is the function we'll actually export for building our configuration.
 -- Since it wraps all our efforts to read information from the command line, and
 -- the file, before combining it all and returning the required information.
+
+-- Remember that we want the command line configuration to take precedence over
+-- the File configuration, so if we think about combining each of our ``Conf``
+-- records. By now we should be able to write something like this:
+--
+-- ``defaults <> file <> commandLine``
+--
 parseOptions
   :: FilePath
   -> IO (Either ConfigError Conf)
-parseOptions fp =
-  let mkCfg cli file = makeConfig (defaultConf <> file <> cli)
-  in do
-    cli' <- commandLineParser
-    ( >>= mkCfg cli' ) <$> parseJSONConfigFile fp
+parseOptions =
+  -- Parse the options from the config file: "files/appconfig.json"
+  -- Parse the options from the commandline using 'commandLineParser'
+  -- Combine these with the default configuration 'defaultConf'
+  -- Return the final configuration value
+  error "parseOptions not implemented"
