+++ src/Level06/Conf/File.hs
@@ -1,3 +1,4 @@
+{-# LANGUAGE OverloadedStrings #-}
 module Level06.Conf.File where
 
 import           Data.ByteString.Lazy       (ByteString)
@@ -10,36 +11,42 @@ import           Data.Monoid                (Last (Last))
 
 import           Control.Exception          (try)
 
-import           Data.Aeson                 (FromJSON, Object, (.:))
+import           Data.Aeson                 (FromJSON, Object)
 
 import qualified Data.Aeson                 as Aeson
-import qualified Data.Aeson.Types           as Aeson
 
-import           Level06.Types             (ConfigError (..), PartialConf)
-
+import           Level06.Types             (ConfigError,
+                                             PartialConf (PartialConf))
 -- Doctest setup section
 -- $setup
 -- >>> :set -XOverloadedStrings
+
+-- | File Parsing
+
+-- We're trying to avoid complications when selecting a configuration file
+-- package from Hackage. We'll use an encoding that you are probably familiar
+-- with, for better or worse, and write a small parser to pull out the bits we
+-- need. The package we're using is the ``aeson`` package to parse some JSON and
+-- we'll pick the bits off the Object.
 
 -- | Update these tests when you've completed this function.
 --
 -- | readConfFile
 -- >>> readConfFile "badFileName.no"
--- Left (ConfigFileReadError badFileName.no: openBinaryFile: does not exist (No such file or directory))
--- >>> readConfFile "test.json"
--- Right "{\"foo\":33}\n"
+-- Left (undefined "badFileName.no: openBinaryFile: does not exist (No such file or directory)")
+-- >>> readConfFile "files/test.json"
+-- Right "{\n  \"foo\": 33\n}\n"
 --
 readConfFile
   :: FilePath
   -> IO ( Either ConfigError ByteString )
-readConfFile fp =
-  first ConfigFileReadError <$> try (LBS.readFile fp)
+readConfFile =
+  error "readConfFile not implemented"
 
 -- Construct the function that will take a ``FilePath``, read it in, decode it,
 -- and construct our ``PartialConf``.
 parseJSONConfigFile
   :: FilePath
   -> IO ( Either ConfigError PartialConf )
-parseJSONConfigFile fp =
-  (first JSONDecodeError . Aeson.eitherDecode =<<) <$> readConfFile fp
-
+parseJSONConfigFile =
+  error "parseJSONConfigFile not implemented"
