+++ src/Level06/README.md
@@ -1,12 +1,33 @@
-# Level 06
+# Level 05
 
-This exercise involves introducing a monad transformer (ReaderT) to our
-application. The use of the ReaderT in our application is motivated in part by
-the expansion of the number of inputs to the various functions involved in
-handling different requests.
+In this exercise we build some configuration capabilities into our application.
 
-Since nearly any given request handler will likely require access to either the
-general application configuration or the database, it is tedious to have to pass
-the information in manually in every instance.
+This exercise will require a combination of building the right types to guide
+your development, plus consulting plenty of documentation to leverage the chosen
+packages. There may also be, depending on your level of interest, some external
+reading for later as well.
 
-Start in ``src/Level06/AppM.hs``.
+The steps for this level:
+1) ``src/Level06/Types.hs``
+2) ``src/Level06/Conf/File.hs``
+3) ``src/Level06/Conf.hs``
+4) ``src/Level06/Main.hs``
+
+The packages we will use for this are:
+
+- [Aeson](http://hackage.haskell.org/package/aeson)
+- [Optparse Applicative](http://hackage.haskell.org/package/optparse-applicative)
+
+#### Aside: Tool Introduction - doctest
+
+This level utilises the [doctest](https://hackage.haskell.org/package/doctest)
+tool to help us ensure our functions comply with some quick tests that are
+written as comments in the source file. This is a port of the same technology
+that exists in Python.
+
+You can see the new entry in the Cabal file as another ``test-suite``. The
+``doctests.hs`` lists the files that have doctests that we want to run. The
+``src/Level06/Conf/File.hs`` file contains some tests that you need to update
+as part of the level.
+
+For details on running and writing doctests, refer to the documentation.
