+++ src/Level06/Types/Error.hs
@@ -2,8 +2,8 @@
 
 module Level06.Types.Error (Error(..), nonEmptyText) where
 
-import Data.Text (Text)
-import Database.SQLite.SimpleErrors.Types (SQLiteResponse)
+import           Data.Text                          (Text)
+import           Database.SQLite.SimpleErrors.Types (SQLiteResponse)
 
 data Error
   = UnknownRoute
