+++ src/Level07/Conf/File.hs
@@ -3,11 +3,17 @@ module Level07.Conf.File where
 import           Data.ByteString.Lazy       (ByteString)
 import qualified Data.ByteString.Lazy.Char8 as LBS
 
+import           Data.Text                  (Text)
+
 import           Data.Bifunctor             (first)
+import           Data.Monoid                (Last (Last))
 
 import           Control.Exception          (try)
 
+import           Data.Aeson                 (FromJSON, Object, (.:))
+
 import qualified Data.Aeson                 as Aeson
+import qualified Data.Aeson.Types           as Aeson
 
 import           Level07.Types             (ConfigError (..), PartialConf)
 
@@ -20,7 +26,7 @@ import           Level07.Types             (ConfigError (..), PartialConf)
 -- | readConfFile
 -- >>> readConfFile "badFileName.no"
 -- Left (ConfigFileReadError badFileName.no: openBinaryFile: does not exist (No such file or directory))
--- >>> readConfFile "test.json"
+-- >>> readConfFile "files/test.json"
 -- Right "{\"foo\":33}\n"
 --
 readConfFile
@@ -36,4 +42,3 @@ parseJSONConfigFile
   -> IO ( Either ConfigError PartialConf )
 parseJSONConfigFile fp =
   (first JSONDecodeError . Aeson.eitherDecode =<<) <$> readConfFile fp
-
