--- src/Level07/appconfig.json
@@ -1,6 +1,0 @@
-{
-  "port": 3000,
-  "helloMsg": "Functional Programming is neat.",
-  "tableName": "comments",
-  "dbName": "firstapp"
-}
