--- src/Level07/test.json
@@ -1,1 +1,0 @@
-{"foo":33}
