+++ tests/Level07Tests.hs
@@ -6,9 +6,7 @@ module Level07Tests
 
 import           Control.Monad.Reader (ask, reader)
 
-import           Data.Monoid          ((<>))
-
-import           Data.String          (IsString)
+import           Control.Monad        (join)
 
 import           Test.Hspec
 import           Test.Hspec.Wai
@@ -32,16 +30,9 @@ doctests =
 
 unitTests :: IO ()
 unitTests = do
-  let
-    dieWith :: Show a => a -> IO ()
-    dieWith m = print m >> Exit.exitFailure
+  let dieWith m = print m >> Exit.exitFailure
 
-    -- This helps keep the string polymorphic so we can use it in both
-    -- ByteString and Text forms in this file, without having to run encoding
-    -- functions. The compiler takes care of it for us.
-    testTopic :: IsString s => s
-    testTopic = "fudge"
-
+  -- Keeping everything in sync with out larger application changes.
   reqsE <- Main.prepareAppReqs
   case reqsE of
 
@@ -51,35 +42,39 @@ unitTests = do
       let app' = pure ( Main.app env )
 
           flushTopic :: IO ()
-          flushTopic = (either dieWith pure =<<)
-            $ AppM.runAppM ( do
-                t <- AppM.liftEither $ Types.mkTopic "fudge"
-                DB.deleteTopic t ) env
+          flushTopic = either ( liftIO . dieWith ) pure =<< AppM.runAppM
+            (AppM.liftEither =<< traverse DB.deleteTopic ( Types.mkTopic "fudge" ))
+            env
+
+      -- We can't run the tests for our AppM in the same stage as our
+      -- application, because of the use of the 'with' function. As it expects
+      -- to be able to execute our tests by applying it to our 'Application'.
+      hspec $ appMTests env
 
       -- Run the tests with a DB topic flush between each spec
       hspec . with ( flushTopic >> app' ) $ do
-        -- Save us a bit of repetition
-        let pOST = post ( "/" <> testTopic <> "/add" )
 
         -- AddRq Spec
         describe "POST /topic/add" $ do
+
           it "Should return 200 with well formed request" $
-            pOST "Is super tasty." `shouldRespondWith` "Success"
+            post "/fudge/add" "Fred" `shouldRespondWith` "Success"
 
           it "Should 400 on empty input" $
-            pOST "" `shouldRespondWith` 400
+            post "/fudge/add" "" `shouldRespondWith` 400
 
         -- ViewRq Spec
         describe "GET /topic/view" $
           it "Should return 200 with content" $ do
-            _ <- pOST "Is super tasty."
-            get ( "/" <> testTopic <> "/view" ) `shouldRespondWith` 200
+            post "/fudge/add" "Is super tasty."
+            get "/fudge/view" `shouldRespondWith` 200
 
         -- ListRq Spec
         describe "GET /list" $
           it "Should return 200 with content" $ do
-            _ <- pOST "Is super tasty."
+            post "/fudge/add" "Is super tasty."
             get "/list" `shouldRespondWith` "[\"fudge\"]"
+
 
 -- These tests ensure that our AppM will do we want it to, with respect to the
 -- behaviour of 'ask', 'reader', and use in a Monad.
@@ -88,8 +83,7 @@ appMTests env = describe "AppM Tests" $ do
 
   it "ask should retrieve the Env" $ do
     r <- AppM.runAppM ask env
-    let cfg = AppM.envConfig <$> r
-    ( cfg == ( Right $ AppM.envConfig env  )) `shouldBe` True
+    ( (AppM.envConfig <$> r) == (Right $ AppM.envConfig env) ) `shouldBe` True
 
   it "reader should run a function on the Env" $ do
     let getDBfilepath = Types.dbFilePath . AppM.envConfig
@@ -102,4 +96,4 @@ appMTests env = describe "AppM Tests" $ do
           e <- ask
           AppM.envLoggingFn e "In a test!"
     r <- AppM.runAppM fn env
-    r `shouldBe` ( Right () )
+    r `shouldBe` (Right ())
